+++ package.json
@@ -20,13 +20,17 @@
     "react-dom": "17.0.2",
     "react-is": "^17.0.2",
     "react-redux": "^7.2.4",
+    "redux-devtools-extension": "^2.13.9",
+    "redux-thunk": "^2.3.0",
     "styled-components": "^5.3.0"
   },
   "devDependencies": {
     "@material-ui/core": "^4.12.2",
     "@types/react": "17.0.11",
     "@types/react-datepicker": "^4.1.1",
-    "@types/react-redux": "^7.1.16",
+    "@types/react-redux": "^7.1.18",
+    "@types/redux": "^3.6.0",
+    "@types/redux-thunk": "^2.1.0",
     "@types/styled-components": "^5.1.10",
     "eslint": "7.29.0",
     "eslint-config-next": "11.0.0",
